llm_chaining/requirements.txt
openai>=1.12.0
anthropic>=0.18.1
python-dotenv>=1.0.1
tqdm>=4.66.2
asyncio>=3.4.3 